feat(init): refactori code
