+++ lua/gen/utils/trim_table.lua
@@ -1,0 +1,17 @@
+local function trim_table(tbl)
+    local function is_whitespace(str)
+        return str:match("^%s*$") ~= nil
+    end
+
+    while #tbl > 0 and (tbl[1] == "" or is_whitespace(tbl[1])) do
+        table.remove(tbl, 1)
+    end
+
+    while #tbl > 0 and (tbl[#tbl] == "" or is_whitespace(tbl[#tbl])) do
+        table.remove(tbl, #tbl)
+    end
+
+    return tbl
+end
+
+return trim_table
