+++ lua/gen/window/get.lua
@@ -1,0 +1,34 @@
+local function get_window_options(opts)
+    local cursor = vim.api.nvim_win_get_cursor(0)
+    local new_win_width = vim.api.nvim_win_get_width(0)
+    local win_height = vim.api.nvim_win_get_height(0)
+
+    local middle_row = win_height / 2
+
+    local new_win_height = math.floor(win_height / 2)
+    local new_win_row
+    if cursor[1] <= middle_row then
+        new_win_row = 5
+    else
+        new_win_row = -5 - new_win_height
+    end
+
+    local result = {
+        relative = "cursor",
+        width = new_win_width,
+        height = new_win_height,
+        row = new_win_row,
+        col = 0,
+        style = "minimal",
+        border = "rounded",
+    }
+
+    local version = vim.version()
+    if version.major > 0 or version.minor >= 10 then
+        result.hide = opts.hidden
+    end
+
+    return result
+end
+
+return get_window_options
